requirements_2.txt
pandas
numpy
faker
sqlalchemy
psycopg2-binary
python-dotenv
selenium
streamlit
plotly_express